Create Drafr 2
Changed the print statements:

print("Choose a letter from the list to correspond with the RGB and HEX color you want.")
print("R equals red", "B equals blue", "Y equals yellow", "O equals orange", "G equals green", "V equals violet")

To comments:

# Choose a letter from the list to correspond with the RGB and HEX color you want.
# R equals red", "B equals blue", "Y equals yellow", "O equals orange", "G equals green", "V equals violet.

The reason why I couldn't get the hex value is because I kept typing in the color in Upper case 

ex. Red. 

When I type the color "Red", it will give me the rgb value for that color because that's how the color is written in the rgb list. Now if you type in "red" 
(all lowercase then the hex value for red will appear. 

So aside from that, my code is done and complete. Just run the code a couple times and I need  try to add the if statements and see if it works, since it 

was originally part of my code plus its on my final project plan that I handed in earlier.